refactor(discover): move components to their own folder
